containerutils: Move package outside gadgettracermanager
Moving out of gadgettracer manager because it is also used by other packages.
