+++ pkg/container-utils/containerd/containerd.go
@@ -17,7 +17,7 @@ package containerd
 import (
 	"time"
 
-	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/gadgettracermanager/containerutils/runtime-client"
+	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/container-utils/runtime-client"
 )
 
 const (
