+++ pkg/container-utils/crio/crio.go
@@ -17,7 +17,7 @@ package crio
 import (
 	"time"
 
-	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/gadgettracermanager/containerutils/runtime-client"
+	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/container-utils/runtime-client"
 )
 
 const (
