+++ pkg/container-utils/docker/docker.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/docker/docker/client"
-	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/gadgettracermanager/containerutils/runtime-client"
+	runtimeclient "github.com/kinvolk/inspektor-gadget/pkg/container-utils/runtime-client"
 )
 
 const (
